* Fixed Riding skill, after [5982]. * Fixed QueryResultPostgre::ConvertNativeType(). Coding Style :p
git-svn-id: http://mangos.svn.sourceforge.net/svnroot/mangos@5992 26674162-90ff-0310-a85b-b64107c8dee9

+++ trunk/src/game/UnitEvents.h
@@ -73,7 +73,7 @@ class MANGOS_DLL_SPEC UnitBaseEvent
         uint32 iType;
     public:
         UnitBaseEvent(uint32 pType) { iType = pType; }
-        uint16 getType() const { return iType; }
+        uint32 getType() const { return iType; }
         bool matchesTypeMask(uint32 pMask) const { return iType & pMask; }
 
         void setType(uint32 pType) { iType = pType; }
